Image_Classification_Example/requirements.txt
tensorflow
numpy
keras
pillow
matplotlib
opencv-python
